feat: move method add as a AddService
move method add as a AddService
