+++ task_list/app.py
@@ -1,5 +1,5 @@
-from task_list.application.domain.model import ProjectName, TaskId, ToDoList, ToDoListId
-from task_list.application.domain.service import HelpService, ShowService
+from task_list.application.domain.model import TaskId, ToDoList, ToDoListId
+from task_list.application.domain.service import AddService, HelpService, ShowService
 from task_list.console import Console
 
 
@@ -25,7 +25,7 @@ class TaskList:
         if command == "show":
             ShowService(todo_list=self.todo_list, console=self.console).show()
         elif command == "add":
-            self.add(command_rest[1])
+            AddService(todo_list=self.todo_list, console=self.console).add(command_rest[1])
         elif command == "check":
             self.check(command_rest[1])
         elif command == "uncheck":
@@ -34,26 +34,6 @@ class TaskList:
             HelpService(console=self.console).help()
         else:
             self.error(command)
-
-    def add(self, command_line: str) -> None:
-        sub_command_rest = command_line.split(" ", 1)
-        sub_command = sub_command_rest[0]
-        if sub_command == "project":
-            self.add_project(ProjectName(value=sub_command_rest[1]))
-        elif sub_command == "task":
-            project_task = sub_command_rest[1].split(" ", 1)
-            self.add_task(ProjectName(value=project_task[0]), project_task[1])
-
-    def add_project(self, project_name: ProjectName) -> None:
-        self.todo_list.add_project(project_name)
-
-    def add_task(self, project_name: ProjectName, description: str) -> None:
-        project = self.todo_list.get_project(project_name)
-        if project is None:
-            self.console.print(f"Could not find a project with the name {project_name}.")
-            self.console.print()
-            return
-        self.todo_list.add_task(project_name=project_name, description=description)
 
     def check(self, task_id: str) -> None:
         self.todo_list.set_done(task_id=TaskId(value=task_id), done=True)
