+++ task_list/application/domain/service/__init__.py
@@ -1,4 +1,5 @@
+from task_list.application.domain.service.add import AddService
 from task_list.application.domain.service.help import HelpService
 from task_list.application.domain.service.show import ShowService
 
-__all__ = ["HelpService", "ShowService"]
+__all__ = ["AddService", "HelpService", "ShowService"]
